// only-doc-vibe-main/src/shared/api/cms/cms-schema.ts
export interface components {
  schemas: {
// ...
    SectionsHeroSectionComponent: {
      id: number;
      /** @enum {string} */
      __component: "sections.hero-section";
      internal_title: string;
      title: string;
      subtitle: string;
      drop_zone: components["schemas"]["WidgetDropZoneComponent"];
      services_group: components["schemas"]["ServiceServicesGroupComponent"];
      upload_processing_modal: components["schemas"]["ModalsUploadProcessingModalComponent"];
      convert_modal: components["schemas"]["ModalsConvertModalComponent"];
    };
// ...
    ServiceServiceCategoryComponent: {
      id: number;
      title: string;
      /** @enum {string} */
      category_id: "edit-pdf" | "convert-from-pdf" | "convert-to-pdf" | "image-tools" | "ai-tools";
      category_icon: {
        id: number;
        documentId: string;
        name: string;
        alternativeText: string;
        caption: string;
        width: number;
        height: number;
        formats: unknown;
        hash: string;
        ext: string;
        mime: string;
        /** Format: float */
        size: number;
        url: string;
        previewUrl: string;
        provider: string;
        provider_metadata: unknown;
        related: {
            id: number;
            documentId: string;
          }[];
        folder: {
          id: number;
          documentId: string;
        };
        folderPath: string;
        /** Format: date-time */
        createdAt: string;
        /** Format: date-time */
        updatedAt: string;
        /** Format: date-time */
        publishedAt: string;
        createdBy: {
          id: number;
          documentId: string;
        };
        updatedBy: {
          id: number;
          documentId: string;
        };
        locale: string;
        localizations: {
            id: number;
            documentId: string;
          }[];
      };
      service_item: components["schemas"]["ServiceServiceTypeComponent"][];
    };
    SectionsServicesComponent: {
      id: number;
      /** @enum {string} */
      __component: "sections.services";
      internal_title: string;
      category_item: components["schemas"]["ServiceServiceCategoryComponent"][];
      feature_highlights: components["schemas"]["CommonImageItemComponent"][];
      show_more_button: string;
    };
// ...
    GlobalBlock: {
      id: number;
      documentId: string;
      title: string;
      widgets: (components["schemas"]["SectionsHeroSectionComponent"] | components["schemas"]["SectionsServicesComponent"] | components["schemas"]["GlobalTrustpilotReviewsComponent"] | components["schemas"]["SectionsAchivementsComponent"] | components["schemas"]["SectionsTrustComponent"])[];
      /** Format: date-time */
      createdAt: string;
      /** Format: date-time */
      updatedAt: string;
      /** Format: date-time */
      publishedAt: string;
      createdBy: {
        id: number;
        documentId: string;
      };
      updatedBy: {
        id: number;
        documentId: string;
      };
      locale: string;
      localizations: {
          id: number;
          documentId: string;
        }[];
    };
    GlobalBlockResponse: {
      data: components["schemas"]["GlobalBlock"];
      meta: Record<string, never>;
    };
// ...
    MainPageResponse: {
      data: components["schemas"]["MainPage"];
      meta: Record<string, never>;
    };
// ...
    NavbarResponse: {
      data: components["schemas"]["Navbar"];
      meta: Record<string, never>;
    };
// ...
  };
// ...
}